Performance improvement: filtering by label
Having to splice the label directly into the query string, since we
can't pass it as a parameter, so I need to make sure that we're not
running into a Bobby Tables scenario - their parser seems to handle it
well though.

Even the straightforward insert test takes < 50% of the time to run
after filtering by labels, and that's considering we don't have indexes
on all labels right now.