Fixed error logging
So this one’s a bit of a bitch to test, since I can’t make this raise an
exception when attempting to get something from TVTropes, and the
function depends on having their URL as a base.

Therefore no tests. Therefore didn’t notice when refactoring the code in
a hurry that I hadn’t made a function for dealing with one URL at a time.

Keeping error as the error message, creating general hasError property
so that we can index it.