Re-wrote page import as a merge
I don't get to do exactly the same as before yet, since I can pass only
one set of parameters for the list to link to (doesn't seem like I can
pass a map).  Therefore when I'm creating the nodes to link to I don't
get to save the URL.

This means that query-nodes-to-crawl may not have a URL. In this case we
construct one out of the known base path that we are crawling, given that
we know it.

I still want to cherry-pick other changes from a different branch, but
doing an intermediate commit first.

Still keeping around the old node create/merge functions, will prune them
in the future.