Significant node network query performance improvement
We don't really need to filter to get just a few of the related nodes.
This saves us having to do an extra query then pass a bunch of node
identifiers as parameters.

Improved query so that both the from and to nodes are filtered when
we want to eliminate links to volume outliers, like nodes with too many
incoming links. Previously we were only filtering on the incoming link
count for the nodes being linked _to_.