Parsing the actual trope items
Going to try my hand at a different visualization, based around the trope
item text.  Added functions to retrieve the trope list items, as well as
tests.

Emitted trope text is wrapped in <span> instead of <li>.
We'll use this text on the front end to embed it directly into lists and
spans, so it's better if it's delivered as spans so we can embed them
anywhere.