Changing hiccup element processing to a composed list of functions
Need to a processor which applies to _all_ elements, since any random
one could have a 'style' that is a string.  This lets me just treat
all elements the same way without getting an ever-growing condp, and
is useful in case we get another filter that needs to be generally
applied.