Trope references sent by /api/tropes now include the codes linked to
We already have the url->code codebase on the backend, so handling the
parsing there instead of doing it on the front end. It requires us to
transmit a bit more data (+10% for Samurai Flamenco) but that won't be a
big deal once gzipped, and saves us some extra, possibly messy front end
processing.