Displaying a graph of the relationships between the liked references
I decided to remove the option to double-click to go into a reference
neighborhood, since we don't actually get that many on the exploration
set and it was cluttering up the code.

Currently passing the code list as part of the GET request, since the
json parser can only retrieve a URL. I could get the data and assign it
myself, which would allow us to keep transit+json, but this is a quick
implementation to experiment

Double-clicking on a node in the graph loads that article.

TODO: New API needs tests once I decide if I'll keep it.